refactor: replace #define with constexpr and long to int8_t

+++ dz1.8 (SMART ALARM)/alarm/include/Button.h
@@ -6,11 +6,11 @@ class Button {
   private:
     const uint8_t pin_;
 
-    unsigned long debounceMs_ = 0;
-    unsigned long clickMaxMs_ = 0;
+    uint32_t debounceMs_ = 0;
+    uint32_t clickMaxMs_ = 0;
 
-    unsigned long lastReadingChangeTime_ = 0;
-    unsigned long pressedAt_ = 0;
+    uint32_t lastReadingChangeTime_ = 0;
+    uint32_t pressedAt_ = 0;
 
     bool pressed_ = false;
     bool lastReading_ = false;
@@ -19,15 +19,15 @@ class Button {
   public:
     Button(
         const uint8_t pin, 
-        unsigned long debounceMs,
-        unsigned long clickMaxMs
+        uint32_t debounceMs,
+        uint32_t clickMaxMs
     );
 
     void init();
-    void update(unsigned long time);
+    void update(uint32_t time);
 
     bool isPressed() const;
     bool wasClicked() const;
 
-    unsigned long heldFor(unsigned long time) const;
+    uint32_t heldFor(uint32_t time) const;
 };
