+++ dz1.8 (SMART ALARM)/alarm/include/LED.h
@@ -6,7 +6,7 @@ class LED {
   private:
     const uint8_t pin_;
     bool state_ = LOW;
-    unsigned long lastBlinkTime_ = 0;
+    uint32_t lastBlinkTime_ = 0;
   
   public:
     explicit LED(const uint8_t pin);
@@ -17,15 +17,7 @@ class LED {
     void turnOn();
     void turnOff();
     void toggle();
-    // void lightBySound(bool sound);
-    // void updateLastBlinkTime(unsigned long time);
 
-    // bool getState() const;
-    // unsigned long getLastBlinkTime() const;
-
-    void resetBlink(unsigned long time);
-    void blink(unsigned long time, unsigned long intervalMs);
-
-    // bool isOn() const;
-    // bool isOff() const;
+    void resetBlink(uint32_t time);
+    void blink(uint32_t time, uint32_t intervalMs);
 };
